Simplify more
